% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{adaptive_bbox}
\alias{adaptive_bbox}
\title{Calculate adaptive bounding box from population coordinates}
\usage{
adaptive_bbox(pop_sf, padding = 0.12)
}
\arguments{
\item{pop_sf}{Population sf object}

\item{padding}{Percentage padding around points (default: 15%)}
}
\value{
Bounding box (xmin, ymin, xmax, ymax)
}
\description{
Calculate adaptive bounding box from population coordinates
}
