% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{create_base_map}
\alias{create_base_map}
\title{Create reusable base map template}
\usage{
create_base_map(spatial_data, bbox)
}
\arguments{
\item{spatial_data}{Spatial data list from load_basemap_data}

\item{bbox}{Bounding box}
}
\value{
tmap object
}
\description{
Create reusable base map template
}
