% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{extract_cv_error}
\alias{extract_cv_error}
\title{Extract CV error from ADMIXTURE log file}
\usage{
extract_cv_error(log_file)
}
\arguments{
\item{log_file}{Path to .out file}
}
\value{
Data frame with K and CV_error
}
\description{
Extract CV error from ADMIXTURE log file
}
