% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{generate_spatial_map}
\alias{generate_spatial_map}
\title{Generate spatial admixture-type visualization components for mapping}
\usage{
generate_spatial_map(admix_df, popmap, popcoords, k_value)
}
\arguments{
\item{admix_df}{Admixture matrix data frame}

\item{popmap}{Population map data frame}

\item{popcoords}{Population coordinates data frame}

\item{k_value}{Number of clusters (K)}
}
\value{
List with population SF object and pie chart grobs
}
\description{
Generate spatial admixture-type visualization components for mapping
}
