% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{get_osm_water_data}
\alias{get_osm_water_data}
\title{Spatial Data Processing Functions}
\usage{
get_osm_water_data(bbox_poly, osm_cache_dir)
}
\arguments{
\item{bbox_poly}{The bounding box polygon}

\item{osm_cache_dir}{Cache directory for OSM data}
}
\value{
List containing rivers, lakes_poly, and lakes_multi
}
\description{
This module contains functions for processing spatial data and creating
geographic visualizations of ADMIXTURE results.
Get OSM water data with caching
}
