% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{get_script_dir}
\alias{get_script_dir}
\title{Command-Line Interface Functions}
\usage{
get_script_dir()
}
\value{
The directory of the R script
}
\description{
This module contains functions for handling command-line arguments
and executing the pipeline from the command line.
Get the directory of the currently running R script
}
