% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{population_pie_chart}
\alias{population_pie_chart}
\title{Create a pie chart for population admixture}
\usage{
population_pie_chart(proportions, colors = NULL)
}
\arguments{
\item{proportions}{Named vector of admixture proportions}

\item{colors}{Vector of colors for clusters}
}
\value{
ggplot object of pie chart
}
\description{
Create a pie chart for population admixture
}
