% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{process_q_files}
\alias{process_q_files}
\title{Process Q files (single or batch)}
\usage{
process_q_files(input_dir_or_qfile, sample_ids, is_single = FALSE)
}
\arguments{
\item{input_dir_or_qfile}{Directory or single file path}

\item{sample_ids}{Sample IDs}

\item{is_single}{Logical: single file mode?}
}
\value{
List of Q matrices
}
\description{
Process Q files (single or batch)
}
