% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{read_q_file}
\alias{read_q_file}
\title{Read Q file, detect K from columns, validate row count}
\usage{
read_q_file(q_path, sample_ids)
}
\arguments{
\item{q_path}{Path to .Q file}

\item{sample_ids}{Vector of sample IDs from FAM}
}
\value{
Q matrix or NULL if K==1
}
\description{
Read Q file, detect K from columns, validate row count
}
