% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{save_genomics_plot}
\alias{save_genomics_plot}
\title{Save plot with standardized settings}
\usage{
save_genomics_plot(plot, filename, width = 8, height = 6, dpi = 300)
}
\arguments{
\item{plot}{ggplot object}

\item{filename}{Output filename}

\item{width}{Plot width in inches}

\item{height}{Plot height in inches}

\item{dpi}{Resolution}
}
\description{
Save plot with standardized settings
}
